rename opencode on the frontend
