+++ backend/src/executors/charm_opencode.rs
