+++ backend/src/executors/mod.rs
@@ -1,15 +1,15 @@
 pub mod amp;
+pub mod charm_opencode;
 pub mod claude;
 pub mod dev_server;
 pub mod echo;
 pub mod gemini;
-pub mod opencode;
 pub mod setup_script;
 
 pub use amp::{AmpExecutor, AmpFollowupExecutor};
+pub use charm_opencode::{CharmOpencodeExecutor, CharmOpencodeFollowupExecutor};
 pub use claude::{ClaudeExecutor, ClaudeFollowupExecutor};
 pub use dev_server::DevServerExecutor;
 pub use echo::EchoExecutor;
 pub use gemini::{GeminiExecutor, GeminiFollowupExecutor};
-pub use opencode::{CharmOpencodeExecutor, CharmOpencodeFollowupExecutor};
 pub use setup_script::SetupScriptExecutor;
