+++ frontend/src/components/tasks/TaskDetailsToolbar.tsx
@@ -21,7 +21,7 @@ const availableExecutors = [
   { id: 'claude', name: 'Claude' },
   { id: 'amp', name: 'Amp' },
   { id: 'gemini', name: 'Gemini' },
-  { id: 'opencode', name: 'OpenCode' },
+  { id: 'charmopencode', name: 'Charm/Opencode' },
 ];
 
 function TaskDetailsToolbar() {
